BLZT-43: split AnnotationBringBeanRegistry into two part ConfigurationBeanRegistrar & BeanCreator.
