+++ core/src/main/java/com/bobocode/bring/core/context/impl/AnnotationBringBeanRegistry.java
@@ -1,45 +1,35 @@
 package com.bobocode.bring.core.context.impl;
 
-import com.bobocode.bring.core.anotation.Autowired;
-import com.bobocode.bring.core.anotation.Value;
 import com.bobocode.bring.core.context.BeanDefinitionRegistry;
 import com.bobocode.bring.core.context.BeanRegistry;
 import com.bobocode.bring.core.context.scaner.ClassPathScannerFactory;
-import com.bobocode.bring.core.context.type.ParameterValueTypeInjector;
 import com.bobocode.bring.core.domain.BeanDefinition;
 import com.bobocode.bring.core.exception.CyclicBeanException;
-import com.bobocode.bring.core.exception.NoConstructorWithAutowiredAnnotationBeanException;
-import com.bobocode.bring.core.exception.NoSuchBeanException;
-import com.bobocode.bring.core.exception.NoUniqueBeanException;
-import com.bobocode.bring.core.utils.ReflectionUtils;
 import lombok.Getter;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.reflections.Reflections;
 
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-import java.lang.reflect.Parameter;
-import java.util.*;
-import java.util.function.Supplier;
-
-import static com.bobocode.bring.core.utils.ReflectionUtils.isAutowiredSetterMethod;
-import static com.bobocode.bring.core.utils.ReflectionUtils.setField;
+import java.util.HashSet;
+import java.util.Set;
 
 @Slf4j
 public class AnnotationBringBeanRegistry extends DefaultBringBeanFactory implements BeanRegistry, BeanDefinitionRegistry {
+    @Getter
     private final ClassPathScannerFactory classPathScannerFactory;
     protected final Set<Class<?>> beansToCreate;
     private final Set<String> currentlyCreatingBeans = new HashSet<>();
 
     @Getter
     private final Reflections reflections;
+    private final ConfigurationBeanRegistrar configurationBeanRegistrar;
+    private final BeanCreator beanCreator;
 
     public AnnotationBringBeanRegistry(Reflections reflections) {
         this.reflections = reflections;
         this.classPathScannerFactory = new ClassPathScannerFactory(reflections);
         this.beansToCreate = classPathScannerFactory.getBeansToCreate();
+        this.configurationBeanRegistrar = new ConfigurationBeanRegistrar(this);
+        this.beanCreator = new BeanCreator(this);
     }
 
     @Override
@@ -60,232 +50,17 @@ public class AnnotationBringBeanRegistry extends DefaultBringBeanFactory impleme
         currentlyCreatingBeans.add(beanName);
 
         if (beanDefinition.isConfigurationBean()) {
-            registerConfigurationBean(beanName, beanDefinition);
+            configurationBeanRegistrar.registerConfigurationBean(beanName, beanDefinition);
         } else {
-            Constructor<?> constructor = findAutowiredConstructor(clazz);
-            createBeanUsingConstructor(constructor, beanName, beanDefinition);
-
-            injectDependencies(clazz, beanName);
+            beanCreator.create(clazz, beanName, beanDefinition);
         }
 
         currentlyCreatingBeans.clear();
-    }
-
-    @SneakyThrows
-    private Object registerConfigurationBean(String beanName, BeanDefinition beanDefinition) {
-        Object configObj = Optional.ofNullable(getSingletonObjects().get(beanDefinition.getFactoryBeanName()))
-                .orElseThrow(() -> {
-                    log.info("Unable to register Bean from Configuration class [{}]: " +
-                            "Configuration class not annotated or is not of Singleton scope.", beanName);
-                    return new NoSuchBeanException(beanDefinition.getBeanClass());
-                });
-        
-        List<String> methodParamNames = ReflectionUtils.getParameterNames(beanDefinition.getMethod());
-
-        List<Object> methodObjs = new ArrayList<>();
-        methodParamNames.forEach(
-                paramName -> {
-                    Object object = getBeanByName(paramName);
-
-                    if (Objects.nonNull(object)) {
-                        methodObjs.add(object);
-                    } else {
-                        BeanDefinition bd = Optional.ofNullable(getBeanDefinitionMap().get(paramName))
-                                .orElseThrow(() -> new NoSuchBeanException(beanDefinition.getBeanClass()));
-                        Object newObj = registerConfigurationBean(bd.getFactoryMethodName(), bd);
-                        methodObjs.add(newObj);
-                    }
-                });
-
-        Supplier<Object> supplier = ReflectionUtils.invokeBeanMethod(beanDefinition.getMethod(), 
-                configObj, methodObjs.toArray());
-        Object bean = supplier.get();
-        
-        if (beanDefinition.isPrototype()) {
-            addPrototypeBean(beanName, supplier);
-        } else {
-            addSingletonBean(beanName, bean);
-        }
-
-        return bean;
     }
 
     @Override
     public void registerBeanDefinition(BeanDefinition beanDefinition) {
         String beanName = classPathScannerFactory.resolveBeanName(beanDefinition.getBeanClass());
         addBeanDefinition(beanName, beanDefinition);
-    }
-
-    private Constructor<?> findAutowiredConstructor(Class<?> clazz) {
-        Constructor<?>[] constructors = clazz.getConstructors();
-
-        if (constructors.length == 1) {
-            return constructors[0];
-        }
-
-        return Arrays.stream(constructors)
-                .filter(constructor -> constructor.isAnnotationPresent(Autowired.class))
-                .findFirst()
-                .orElseThrow(() -> new NoConstructorWithAutowiredAnnotationBeanException(
-                        clazz, Arrays.toString(constructors)));
-    }
-
-    private void createBeanUsingConstructor(Constructor<?> constructor, String beanName,
-                                            BeanDefinition beanDefinition) {
-
-        Parameter[] parameters = constructor.getParameters();
-        Object[] dependencies = new Object[parameters.length];
-
-        List<String> parameterNames = ReflectionUtils.getParameterNames(constructor);
-
-        for (int i = 0; i < parameters.length; i++) {
-            Parameter parameter = parameters[i];
-            int index = i;
-            Optional<Object> injectViaProperties = getTypeResolverFactory().getParameterValueTypeInjectors().stream()
-                    .filter(valueType -> valueType.hasAnnotatedWithValue(parameter))
-                    .map(valueType -> {
-                        Object dependencyValue = valueType.setValueToSetter(parameter, classPathScannerFactory.getCreatedBeanAnnotations());
-                        if (dependencyValue instanceof List) {
-                            return injectListDependency((List<Class<?>>) dependencyValue);
-                        }
-                        return dependencyValue;
-                    })
-                    .map(obj -> dependencies[index] = obj)
-                    .findFirst();
-
-            if (injectViaProperties.isEmpty()) {
-                String dependencyBeanName = findBeanNameForArgumentInConstructor(parameter, parameterNames);
-                Object dependencyObject = getOrCreateBean(dependencyBeanName);
-                dependencies[i] = dependencyObject;
-            }
-        }
-
-        Supplier<Object> supplier = ReflectionUtils.createNewInstance(constructor, dependencies, 
-                beanDefinition.getBeanClass(), beanDefinition.isProxy());
-        Object bean = supplier.get();
-
-        if (beanDefinition.isPrototype()) {
-            addPrototypeBean(beanName, supplier);
-        } else {
-            addSingletonBean(beanName, bean);
-        }
-    }
-
-    private String findBeanNameForArgumentInConstructor(Parameter parameter, List<String> constructorParamNames) {
-        Class<?> clazz = parameter.getType();
-
-        List<String> beanNames = clazz.isInterface()
-                ? getTypeToBeanNames().entrySet().stream()
-                    .filter(entry -> clazz.isAssignableFrom(entry.getKey()))
-                    .map(Map.Entry::getValue)
-                    .flatMap(Collection::stream)
-                    .toList()
-                : Optional.ofNullable(getTypeToBeanNames().get(clazz)).orElse(Collections.emptyList());
-
-        if (beanNames.isEmpty()) {
-            throw new NoSuchBeanException(clazz);
-        } else if (beanNames.size() == 1) {
-            return beanNames.get(0);
-        } else {
-            String paramName = constructorParamNames.get(ReflectionUtils.extractParameterPosition(parameter));
-
-            return beanNames.stream()
-                    .filter(name -> name.equalsIgnoreCase(paramName))
-                    .findFirst()
-                    .orElseThrow(() -> new NoUniqueBeanException(clazz, beanNames));
-        }
-    }
-
-    private Object getOrCreateBean(String beanName) {
-        Object existingBean = getBeanByName(beanName);
-
-        if (Objects.nonNull(existingBean)) {
-            return existingBean;
-        }
-
-        BeanDefinition beanDefinition = getBeanDefinitionMap().get(beanName);
-
-        if (Objects.nonNull(beanDefinition)) {
-            registerBean(beanName, beanDefinition);
-        }
-
-        return getBeanByName(beanName);
-    }
-
-    private void injectDependencies(Class<?> clazz, String beanName) {
-        Object bean = getSingletonObjects().get(beanName);
-
-        Arrays.stream(clazz.getDeclaredFields())
-                .filter(field -> field.isAnnotationPresent(Autowired.class) || field.isAnnotationPresent(Value.class))
-                .forEach(field -> injectDependencyViaField(field, bean));
-
-        Arrays.stream(clazz.getDeclaredMethods())
-                .filter(ReflectionUtils::isAutowiredSetterMethod)
-                .forEach(method -> injectDependencyViaMethod(method, bean));
-    }
-
-    @SneakyThrows
-    private void injectDependencyViaMethod(Method method, Object bean) {
-        if (isAutowiredSetterMethod(method)) {
-            Parameter parameter = method.getParameters()[0];
-            Optional<Object> injectViaProperties = getTypeResolverFactory().getParameterValueTypeInjectors().stream()
-                    .filter(valueType -> valueType.hasAnnotatedWithValue(parameter))
-                    .map(valueType -> injectDependencyViaParameter(method, parameter, bean, valueType))
-                    .findFirst();
-
-            if (injectViaProperties.isEmpty()) {
-                Class<?> dependencyType = method.getParameterTypes()[0];
-                String dependencyBeanName = classPathScannerFactory.resolveBeanName(dependencyType);
-                Object dependencyObject = getOrCreateBean(dependencyBeanName);
-                method.invoke(bean, dependencyObject);
-            }
-        }
-    }
-
-    @SneakyThrows
-    private Object injectDependencyViaParameter(Method method, Parameter parameter, Object bean,
-                                                ParameterValueTypeInjector valueType) {
-        Object dependencyValue = valueType.setValueToSetter(parameter, classPathScannerFactory.getCreatedBeanAnnotations());
-        if (dependencyValue instanceof List) {
-            var dependencyObjects = injectListDependency((List<Class<?>>) dependencyValue);
-            method.invoke(bean, dependencyObjects);
-        } else {
-            method.invoke(bean, dependencyValue);
-        }
-        return dependencyValue;
-    }
-
-    @SneakyThrows
-    private void injectDependencyViaField(Field field, Object bean) {
-        Optional<Object> injectViaProperties = getTypeResolverFactory().getFieldValueTypeInjectors().stream()
-                .filter(valueType -> valueType.hasAnnotatedWithValue(field))
-
-                .map(valueType -> {
-                    Object dependencyValue = valueType.setValueToField(field, bean, classPathScannerFactory.getCreatedBeanAnnotations());
-                    if (dependencyValue instanceof List listDependencyValue) {
-                        setField(field, bean, injectListDependency(listDependencyValue));
-                    } else {
-                        setField(field, bean, dependencyValue);
-                    }
-                    return dependencyValue;
-                })
-                .findFirst();
-
-        if (injectViaProperties.isEmpty()) {
-            String dependencyBeanName = classPathScannerFactory.resolveBeanName(field.getType());
-            Object dependencyObject = getOrCreateBean(dependencyBeanName);
-            setField(field, bean, dependencyObject);
-        }
-    }
-
-    private List<Object> injectListDependency(List<Class<?>> value) {
-        List<Object> dependencyObjects = new ArrayList<>();
-        for (var impl : value) {
-            String implBeanName = classPathScannerFactory.resolveBeanName(impl);
-            Object dependecyObject = getOrCreateBean(implBeanName);
-            dependencyObjects.add(dependecyObject);
-        }
-
-        return dependencyObjects;
     }
 }
